+++ src/GitHub_Updater/Bitbucket_API.php
@@ -34,7 +34,7 @@ class Bitbucket_API extends API {
 	 */
 	public function __construct( $type ) {
 		$this->type     = $type;
-		$this->response = $this->get_transient();
+		$this->response = $this->get_repo_cache();
 
 		$this->load_hooks();
 
@@ -80,7 +80,7 @@ class Bitbucket_API extends API {
 			if ( $response ) {
 				$contents = $response->data;
 				$response = $this->get_file_headers( $contents, $this->type->type );
-				$this->set_transient( $file, $response );
+				$this->set_repo_cache( $file, $response );
 			}
 		}
 
@@ -114,7 +114,7 @@ class Bitbucket_API extends API {
 
 			if ( $response ) {
 				$response = $this->parse_tag_response( $response );
-				$this->set_transient( 'tags', $response );
+				$this->set_repo_cache( 'tags', $response );
 			}
 		}
 
@@ -145,7 +145,7 @@ class Bitbucket_API extends API {
 			$content  = $this->get_local_info( $this->type, $changes );
 			if ( $content ) {
 				$response['changes'] = $content;
-				$this->set_transient( 'changes', $response );
+				$this->set_repo_cache( 'changes', $response );
 			} else {
 				$response = false;
 			}
@@ -161,7 +161,7 @@ class Bitbucket_API extends API {
 
 			if ( $response ) {
 				$response = $this->parse_changelog_response( $response );
-				$this->set_transient( 'changes', $response );
+				$this->set_repo_cache( 'changes', $response );
 			}
 		}
 
@@ -218,7 +218,7 @@ class Bitbucket_API extends API {
 			$file     = $response->data;
 			$parser   = new Readme_Parser( $file );
 			$response = $parser->parse_data();
-			$this->set_transient( 'readme', $response );
+			$this->set_repo_cache( 'readme', $response );
 		}
 
 		if ( $this->validate_response( $response ) ) {
@@ -243,7 +243,7 @@ class Bitbucket_API extends API {
 
 			if ( $response ) {
 				$response = $this->parse_meta_response( $response );
-				$this->set_transient( 'meta', $response );
+				$this->set_repo_cache( 'meta', $response );
 			}
 		}
 
@@ -278,7 +278,7 @@ class Bitbucket_API extends API {
 					$branches[ $branch ] = $this->construct_download_link( false, $branch );
 				}
 				$this->type->branches = $branches;
-				$this->set_transient( 'branches', $branches );
+				$this->set_repo_cache( 'branches', $branches );
 
 				return true;
 			}
@@ -372,7 +372,7 @@ class Bitbucket_API extends API {
 					$response->{$locale->language}->version = $this->type->remote_version;
 				}
 
-				$this->set_transient( 'languages', $response );
+				$this->set_repo_cache( 'languages', $response );
 			}
 		}
 		$this->type->language_packs = $response;