+++ src/GitHub_Updater/GitLab_API.php
@@ -41,7 +41,7 @@ class GitLab_API extends API {
 	 */
 	public function __construct( $type ) {
 		$this->type     = $type;
-		$this->response = $this->get_transient();
+		$this->response = $this->get_repo_cache();
 
 		if ( ! isset( self::$options['gitlab_access_token'] ) ) {
 			self::$options['gitlab_access_token'] = null;
@@ -85,7 +85,7 @@ class GitLab_API extends API {
 			if ( $response ) {
 				$contents = base64_decode( $response->content );
 				$response = $this->get_file_headers( $contents, $this->type->type );
-				$this->set_transient( $file, $response );
+				$this->set_repo_cache( $file, $response );
 			}
 		}
 
@@ -120,7 +120,7 @@ class GitLab_API extends API {
 
 			if ( $response ) {
 				$response = $this->parse_tag_response( $response );
-				$this->set_transient( 'tags', $response );
+				$this->set_repo_cache( 'tags', $response );
 			}
 		}
 
@@ -151,7 +151,7 @@ class GitLab_API extends API {
 			$content  = $this->get_local_info( $this->type, $changes );
 			if ( $content ) {
 				$response['changes'] = $content;
-				$this->set_transient( 'changes', $response );
+				$this->set_repo_cache( 'changes', $response );
 			} else {
 				$response = false;
 			}
@@ -164,7 +164,7 @@ class GitLab_API extends API {
 
 			if ( $response ) {
 				$response = $this->parse_changelog_response( $response );
-				$this->set_transient( 'changes', $response );
+				$this->set_repo_cache( 'changes', $response );
 			}
 		}
 
@@ -217,7 +217,7 @@ class GitLab_API extends API {
 			$file     = base64_decode( $response->content );
 			$parser   = new Readme_Parser( $file );
 			$response = $parser->parse_data();
-			$this->set_transient( 'readme', $response );
+			$this->set_repo_cache( 'readme', $response );
 		}
 
 		if ( $this->validate_response( $response ) ) {
@@ -250,8 +250,8 @@ class GitLab_API extends API {
 
 			if ( $response ) {
 				$response = $this->parse_meta_response( $response );
-				$this->set_transient( 'meta', $response );
-				$this->set_transient( 'project', null );
+				$this->set_repo_cache( 'meta', $response );
+				$this->set_repo_cache( 'project', null );
 			}
 		}
 
@@ -288,7 +288,7 @@ class GitLab_API extends API {
 					$branches[ $branch->name ] = $this->construct_download_link( false, $branch->name );
 				}
 				$this->type->branches = $branches;
-				$this->set_transient( 'branches', $branches );
+				$this->set_repo_cache( 'branches', $branches );
 
 				return true;
 			}
@@ -410,7 +410,7 @@ class GitLab_API extends API {
 					$response->{$locale->language}->version = $this->type->remote_version;
 				}
 
-				$this->set_transient( 'languages', $response );
+				$this->set_repo_cache( 'languages', $response );
 			}
 		}
 		$this->type->language_packs = $response;
@@ -515,8 +515,8 @@ class GitLab_API extends API {
 			foreach ( (array) $response as $project ) {
 				if ( $this->type->repo === $project->path ) {
 					$id = $project->id;
-					$this->set_transient( 'project_id', $id );
-					$this->set_transient( 'project', $project );
+					$this->set_repo_cache( 'project_id', $id );
+					$this->set_repo_cache( 'project', $project );
 
 					return $id;
 				}